Adds two buttons to the admin traitor panel (#65688)
About The Pull Request

    Adds "Open VV" to the admin traitor panel, which opens that antagonist's VV panel.
    The "Mark as in/complete" button in the traitor panel now updates color based on whether the objective is completed, instead of whether an admin has marked it to be complete.
        In most cases, it'll end up looking the same / similar as it does currently.
        If the objective is complete without admin intervention, it'll also show up as green
        Some objectives disregard the admin completion value, in which case it will not change color, only text

Why It's Good For The Game

    More complex antags, like lings or heretics, often require that admins view the datum's variables to debug, This makes it easier / faster to locate them.
    People will often ask if their objective is complete or not, and heretics straight up require complete objectives. This makes it easier for admins to see which are complete or not by the code's standards.

Changelog

cl Melbert
admin: Adds "Open VV" to traitor panel antags, which opens that antag's vv panel.
admin: "Mark as in/completed" by objectives in the traitor panel now changes color based on whether the objective is determined to be "complete", instead of whether an admin has marked it to be complete. For some objectives, it might not update color (only text) when an admin marks it as completed - this means the objective disregards whether an admin has marked it as complete.
/cl

(cherry picked from commit 056a5864094ecaa699056683e5ba278774ce9386)