[s] Removes the Gripper Gloves from the Lavaland Syndicate Base (#65948)
Hey there,

These aren't particularly good. Apparently, with these, you are able to _throw_ someone else out of the balcony, have them crawl the rest of the way to shore, and then extinguisher, and then burn meds, and now the highly deadly Syndicate agent with a sniper rifle is free to have "fun" on Lavaland.

This method was discovered quite recently, and it goes against the set policy as well as the design decision made to have the inhabitants of the Lavaland base be able to escape.

(cherry picked from commit cfcfe58d7ec218e9b92ff22e77a65ce60e6aecfb)