Fixes Newscaster Placement Across Three Stations (Click to find out more) (#65992)
Hey there,

So, the changes in the newscaster sprite caused some new weirdness in their placement in, specifically, the AI Satellite. It would block important buttons, like this on Delta:

Or, just be in a totally weird position, like on Meta:

This PR just swaps around some stuff to ensure that everything is accessible to the AI to some level.

(cherry picked from commit d00b29feeff83e1b8757eb02456bdae032ba63ce)