Adds admin logging to supply orders' rare erroneous manifests mechanic (#66001)
Cargo orders currently have a weird undocumented mechanic where there is a small chance for the manifest to generate an incorrect listing for the station name, the order's contents or for the content to go missing altogether. As it isn't a well known feature, plenty of people ahelp it thinking that they have been victims of a bug and as the event isn't logged anywhere, the online staff is inclined to think so if they haven't code dived for it before.

(cherry picked from commit 4006466018fbdb1122a344e7892998b54aaa5830)