Fixes The Lizard's Gas Posters Showing Up On-Station (#66328)
Hey there,

There's been a bug the last four months where the posters made for The Lizard's Gas would show up in random poster spawners. I was aware of it within days of its first sighting, but I elected not to fix it because I thought it was cute with the implication that the crew was going to the gas station and looting the posters, bringing it back to station. Anyways, I've been told to change it, but that's fine. I just turn it into a "fluff" subtype, which shouldn't show up on station or in consoles any more. Very epic.

(cherry picked from commit 2352e654df3db1f98c9262fbe19b95b1d6d3969f)