Fixes Kilostation whiteship (#66468)
Fixes the dimensions on the kilostation whiteship obj so that it spawns the full whiteship. The vars were not correct, making it so the rest of the ship did not spawn.

Tested it on local and the ship fully spawned and moved normally.

(cherry picked from commit 11d7423571f699f1caef115971882c7ca95691f4)