Fixed Icebox's lower two z-levels not being included in the Map Compile action (#66503)
Did you know that you could currently put a bunch of random shit in the lower levels of icebox and the map compile would be none the wiser?

I sure did.

I hate that it's done manually this way, but honestly it's not worth refactoring the whole action to make it work differently.

Ensuring that the lower levels work properly is, in fact, a good thing.

(cherry picked from commit ce0aff7526158133acd1b53bd5d2d9d6ac9578f3)