Places Greater Importance on Justifying Your Changes in the Why Its Good for the Game Section (#66426)
About The Pull Request

This PR:

    Moves the part explaining why you should justify your changes to its own section in contributing.md, and expands on it greatly.
    Adjusts the pull request template to place a further emphasis on arguing for why your changes are good in the Why It's Good For The Game section.

Why It's Good For The Game

Changes that piss people off because of x, y, and z can have merit in spite of that, but that merit should be explained. Period. There is no downside to putting in a reasonable effort to justify why what you're doing will improve the game.

Lazy, low effort explanations for the merits of your changes, wholly unsuited to the level of justification needed to be considered adequate by anyone, should no longer be accepted into our codebase until they are at the very least adequate. This isn't stuff like bugfixes or tiny features no one has any problems with, this is for stuff like #58038 or #62947 or #57831 or #64693 which have either abysmal justification for their changes or are outright refusing any real attempt at a justification relative to their effects.

Proper justification for controversial PRs should be required even if you have explicit support of a maintainer, or maintainers, or a head coder, or multiple head coders, or the entirety of humanity, or God himself before the pr is even made. Because when a PR is merged, the argument for its changes become the codebases argument instead of just the authors. Thus all potentially controversial changes should have a level of effort put into their arguments proportional to their effect on the game.

(cherry picked from commit 1395e53bec927990fe6278c49f5e8dd8cb950631)