Gives atmos techs general engineering access by default (#66751)
Having general access to your department is great for design consistency and allows us to properly map access helpers to reflect what is supposed to be general access and what is not, making it so we do not have to map multiple accesses onto one door when it is not required.

Should also fix Atmospherics Technicians not having access to the Supermatter shiftstart.

(cherry picked from commit 2ef11ea44234feaae20c378419f7b484ef9c9638)