Blacklists holodeck computer from unit_test/create_and_destroy (#66765)
The holodeck runtimes if it's not linked to a holodeck area.
When running integration tests on the default CI map, metastation, this is fine because the one created by create_and_destroy just links to the holodeck on-station. This technically would cause problems in a real game since it's a 2-way link, (create_and_destroy breaks the link to the real on-station computer when it does its test) but no unit tests check that.
Unfortunately, runtime and multiz debug don't have holodecks (for good reason, it's completely unnecessary for testing literally anything but specifically the holodeck) and fail CI as a result.

Since the unit test isn't trying to test for the existence of a well-designed recreation area, I've set unit_test/create_and_destroy to ignore the holodeck computer.

Edit: I've added better mapping error handling for this case in the holodeck. I still don't want a pointless mapping error in my CI either so I'm keeping the ignore.
(cherry picked from commit 129bc94df26309316240f103176a1c297703b61e)