Add silver requirement for surgery tools (#66414)
This adds silver as a requirement to make surgery tools from the autolathe. This idea was taken from #66278 when another person suggested it as an alternative which I liked. I made it so that half the amount of iron needed is the requirement for silver for most surgery items.

10 stacks of silver are needed to make a full set of surgery tools + iron and glass requirements.

(cherry picked from commit e3e3abf690eb6900e83be3560e0d43da48e376b0)