We All Lift Together: fixing an issues
Cut some builidings in a village.
Made a caves on a north of village not so empty.
Removed a "under construction" building near the road, and signs on a road.
Fixed a shroom farm on z3-level.
Placed a roof helpers on a z3-level.
Changed a loot in a secret cave.
Changed a code of paper.
Fixed a stair issue on z2-level (ground level).