We All Lift Together: Mine and Underground Village (#2075)
* We All Lift Together: Mine and Underground Village

Heavily reworked a Mine area. Added a 4 new buildings: Administrative, Cafeteria, Warehouse, Equipment Building.
Reworked ground level of Mines - now there's a few buildings in a caves, with 2nd floor.

Slightly reworked an Underground village: replaced a couple of buildings, added more loot and a few more radscorpions.

* We All Lift Together: fixing an issues

Cut some builidings in a village.
Made a caves on a north of village not so empty.
Removed a "under construction" building near the road, and signs on a road.
Fixed a shroom farm on z3-level.
Placed a roof helpers on a z3-level.
Changed a loot in a secret cave.
Changed a code of paper.
Fixed a stair issue on z2-level (ground level).

* Indentation on Paper.dm

Yeah we webediting. Keep scrolling.

* Adds pickaxes in, touches up loot and fixes some dirs

Co-authored-by: Edward Nashton <eddienigma48@gmail.com>
Co-authored-by: ProfessorPopoff <51188571+ProfessorPopoff@users.noreply.github.com>
Co-authored-by: Professor Popoff <omniderpectional@gmail.com>