Fixes Duplicate APC on IceBoxStation's Prison Wing (#66866)
Hey there, it's on the tin.

I believe the original intent was to have one APC for the Prison Wing, and then another APC for the Prison Wing cells, but I believe an oversight occurred to the point where this APC was not located in the "Safe Prison" area. To rectify this, I put that whole room in that aforementioned area.

This should be a good prerequisite to #66864.

(cherry picked from commit 3acef52c365c8e8220eeffe816ef4d0b4f698706)