Access Helpers Pt. 5/6 - Tram Station (Station Proper) (#66877)
This PR continues to replace door accesses with mapping helpers with the same design standards as #66588, #66753, #66805, and #66871 but this time with TramStation. This does not touch the ~20 different modular maintenance designs which will be done in a future PR.

Standardizes access requirements for the map and reduces changes for inconsistent mapping to happen by use of mapping helpers. Makes access requirements easier to audit.

(cherry picked from commit 1039db78f54db09c2bf4ef08d32116c947a5d981)