Update Time: Power Plant and Clinic (#2087)
* Update Time: Power Plant and Clinic

Remade a second floor of Power Plant and Administrative Building nearby.
Slightly expand a "radioactive waste warehouse" on a 1st floor.
Slightly remade a 1st floors on Power Plant area.

Remade a basement of Clinic.

Add some new stuff on a 2nd floor of Clinic.

* Update Time: small fix

Fixed a missed area on a 2nd floor of a Administrative Building. My bad.

* Update Time: SS13 autististic interaction.

Fixed a stuff. Fuck an interactions.

* Updates shutters in powerplant/hospital and misc fixes/purges

Co-authored-by: Edward Nashton <eddienigma48@gmail.com>
Co-authored-by: Professor Popoff <omniderpectional@gmail.com>